TLE subcommand was implemented

+++ src/runner/types.rs
@@ -2,9 +2,9 @@ use std::time::Duration;
 
 pub trait Compiler {
     fn compile(&self);
-    fn execute(&self, timeout: u64) -> Duration;
+    fn execute(&self, timeout: u32) -> Duration;
 }
 
 pub trait Interpreter {
-    fn execute(&self, timeout: u64) -> Duration;
+    fn execute(&self, timeout: u32) -> Duration;
 }
